test(ci): serialize metrics env budget tests
