+++ tests/grouped/timeseries_remaining/shared.rs
@@ -1,0 +1,34 @@
+use std::sync::{Mutex, MutexGuard, OnceLock};
+
+fn metrics_env_mutex() -> &'static Mutex<()> {
+    static LOCK: OnceLock<Mutex<()>> = OnceLock::new();
+    LOCK.get_or_init(|| Mutex::new(()))
+}
+
+pub fn metrics_env_lock() -> MutexGuard<'static, ()> {
+    metrics_env_mutex()
+        .lock()
+        .unwrap_or_else(|err| err.into_inner())
+}
+
+pub struct EnvGuard {
+    key: &'static str,
+    previous: Option<String>,
+}
+
+impl EnvGuard {
+    pub fn set(key: &'static str, value: &str) -> Self {
+        let previous = std::env::var(key).ok();
+        std::env::set_var(key, value);
+        Self { key, previous }
+    }
+}
+
+impl Drop for EnvGuard {
+    fn drop(&mut self) {
+        match &self.previous {
+            Some(value) => std::env::set_var(self.key, value),
+            None => std::env::remove_var(self.key),
+        }
+    }
+}
